taskLogs/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.7.1
Django==4.1.13
django-allauth==0.61.1
djongo==1.3.6
dnspython==2.6.1
idna==3.6
kafka-python==2.0.2
kombu==5.3.5
oauthlib==3.2.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==1.7.1
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
